refactor: use helper function
